[IMP] doo: Move CSS away + improve userscript
Replace the setTimeout function by a MutationObserver.
The MutationObserver will look if the element we seek is present
in the document. If the MutationObserver finds the element, it will
go through the callback function we define (and inside disconnect the
observer).

Hover on dark mode button is now subject of the URL of the page. Closes #4

Improve regex for localhost.

Make the retrieving of CSS asynchronous.

Add connect option for security.